src/script/ceph-fs-report/requirements.txt
argparse
githubpy 
GitPython
requests_cache
